Make changes necessary to convert rgb images to aron format
When using ndarray_to_aron() to convert rgb images, is_point_cloud()
would return true because the images have fields. To prevent that an
additional condition for a ndarray to be classified as a point cloud
was added. Also the function convert_int8_to_dtype_rgb() to convert
images to an aron format was added.